requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
outcome==1.3.0.post0
playwright==1.47.0
playwright-stealth==1.0.6
propcache==0.2.0
pycparser==2.22
pyee==12.0.0
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
selenium==4.25.0
setuptools==75.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
yarl==1.15.5
